+++ apps/shared/src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './ErrorBoundary';
 export { ToastProvider } from './ToastProvider';
+export { default as Notification } from './Notification';
