+++ apps/shared/src/hooks/index.ts
@@ -1,29 +1,1 @@
-import { useState, ChangeEvent } from 'react';
-
-export interface FormValues {
-  [key: string]: string;
-}
-
-export interface FormErrors {
-  [key: string]: string;
-}
-
-export function useForm<T extends FormValues>(initialValues: T) {
-  const [values, setValues] = useState<T>(initialValues);
-  const [errors, setErrors] = useState<Partial<Record<keyof T, string>>>({});
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const { name, value } = e.target;
-    setValues(prev => ({
-      ...prev,
-      [name]: value
-    }));
-  };
-
-  return {
-    values,
-    errors,
-    handleChange,
-    setErrors,
-  };
-}
+export { useForm } from './useForm';
