+++ apps/shared/src/index.tsx
@@ -1,1 +1,7 @@
-export { ToastProvider, useToast } from './components/ToastProvider';
+export { ToastProvider } from './components/ToastProvider';
+export { useToast } from './components/ToastProvider';
+export { ErrorBoundary } from './components';
+export { Notification } from './components';
+
+export * from './hooks';
+export * from './utils';
